% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mtx}
\alias{example_mtx}
\title{Transcription factor expression matrix}
\format{
A matrix, genes by cells.
}
\usage{
example_mtx
}
\description{
TFs expression matrix sampled from Zeisel A, et al., 2015. Data are normalized
to sequence depth and log2 transformed.
}
\keyword{datasets}
