% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_binding.R
\name{pwm_match}
\alias{pwm_match}
\title{Get TF binding sites}
\usage{
pwm_match(tf_pfm, target_seq, min_score = "85\%")
}
\arguments{
\item{tf_pfm}{position frequency matrix, obtained from MotifDb.}

\item{target_seq}{target gene promoter seq, a DNSstringset object.}

\item{min_score}{min_score for the PWM searching used in \code{matchPWM} (default "85\%").}
}
\value{
The number of pfm binding sites on both postive and negative strands.
}
\description{
Predict the binding sites of a TF on a target sequence by searching pwm.
}
