% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_binding.R
\name{pwm_pred}
\alias{pwm_pred}
\title{Identify mutual bindings}
\usage{
pwm_pred(
  mtx,
  species = "Hsapiens",
  gene_db,
  bsgenome,
  txdb,
  upstream = 1000,
  downstream = 200,
  min_score = "85\%",
  verbose = FALSE
)
}
\arguments{
\item{mtx}{a symmetric TFs connectivity matrix.}

\item{species}{in which species looking for the pfm of TFs, e.g. "Hsapiens" (default), "Mmusculus".}

\item{gene_db}{org.Xx.eg.db annotation converting gene symbol to entrez id, e.g., \code{org.Mm.eg.db::org.Mm.egSYMBOL2EG}.}

\item{bsgenome}{bsgenome of the studied species, e.g., "BSgenome.Hsapiens.UCSC.hg19".}

\item{txdb}{txdb database, e.g., "TxDb.Hsapiens.UCSC.hg19.knownGene" for human.}

\item{upstream}{upstream range of the TSS.}

\item{downstream}{downstream range of the TSS.}

\item{min_score}{min_score for the PWM searching used in \code{matchPWM} (default "85\%").}

\item{verbose}{whether show more info.}
}
\value{
A matrix indicates whether there are TFs binding sites between genes (0 or number of bounded PWMs).
}
\description{
Examine the mutual bindings for each pair of TF in each other's promoter sequences.
}
\details{
The input is the TFs netowrk , predicted from \code{skel_net} or \code{pcn}.
The binding sites of a pair of TFs were predicted on both forward
and reverse strand for available PWMs sepecfied by parameters in a mutual way (i.e., pwms of A on the promter seqs of B, and the other way around).
The numbers are TF PWMs bound on promoter regions (one TF might have multiple PWMs), but not the number of actual binding sites.

The TF position weight matrices were obainted from MotifDb, and binding sites are identified by \code{matchPWM}.
}
