% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_corr.R
\name{skel_net}
\alias{skel_net}
\title{Partial Correlation Network}
\usage{
skel_net(mtx, threshold = "BH", alpha = 0.05)
}
\arguments{
\item{mtx}{single-cell expression matrix (log2 normalized data), genes by cells. Typically genes are transcription factors.}

\item{threshold}{parameter passed to \code{threshold} in \code{qgraph}, could be "sig" without multiple testing correction, or "BH" for fdr.
See \code{qgraph} for detail.}

\item{alpha}{threshold to select significant edges.}
}
\value{
A matrix.
}
\description{
Build a partial correlation network with scRNAseq data.
}
\details{
Partial correlation network based on spearman correlation.
}
