% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_corr.R
\name{pcn}
\alias{pcn}
\title{PCN on large datasets}
\usage{
pcn(mtx, n_cell = NULL, n_draw = 50, ...)
}
\arguments{
\item{mtx}{single-cell expression matrix (log2 normalized data), genes by cells. Typically genes are transcription factors.}

\item{n_cell}{number of cells to sample. If NULL, boostrapping the data and keep unique sets of cells.}

\item{n_draw}{times of sampling.}

\item{...}{additional parameters to \code{skel_net}.}
}
\value{
A matrix.
}
\description{
Build a consensus netork based on partial correlation networks on subsamples of large scRNAseq data.
}
